requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# MT5 Integration
MetaTrader5>=5.0.45

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Technical Analysis
TA-Lib>=0.4.28

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.14.0
joblib>=1.3.0
imbalanced-learn>=0.11.0
optuna>=3.4.0
shap>=0.43.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0
kaleido>=0.2.1

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Reporting
reportlab>=4.0.0
Jinja2>=3.1.2
PyPDF2>=3.0.0

# Monitoring & Logging
loguru>=0.7.0
psutil>=5.9.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
